requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
python-multipart>=0.0.12
